+++ src/main/java/com/devita/common/config/SecurityConfig.java
@@ -2,9 +2,6 @@ package com.devita.common.config;
 
 import com.devita.common.jwt.JwtAuthenticationFilter;
 import com.devita.common.jwt.JwtTokenProvider;
-import com.devita.common.oauth.CustomOAuth2UserService;
-import com.devita.common.oauth.OAuth2LoginSuccessHandler;
-import com.devita.common.oauth.OAuth2LogoutSuccessHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -29,9 +26,6 @@ import java.util.List;
 public class SecurityConfig {
 
     private final JwtTokenProvider jwtTokenProvider;
-    private final CustomOAuth2UserService customOAuth2UserService;
-    private final OAuth2LoginSuccessHandler oAuth2LoginSuccessHandler;
-    private final OAuth2LogoutSuccessHandler oAuth2LogoutSuccessHandler;
 
     @Value("${cors.allowed-origins}")
     private String corsOrigin;
@@ -65,22 +59,6 @@ public class SecurityConfig {
                         .requestMatchers(WHITE_LIST_URL).permitAll()
                         .anyRequest().authenticated()
                 )
-                .oauth2Login(oauth2 -> oauth2
-                        .loginPage("/oauth2/authorization/kakao")
-                        .userInfoEndpoint(userInfo -> userInfo
-                                .userService(customOAuth2UserService)
-                        )
-                        .successHandler(oAuth2LoginSuccessHandler)
-                        .failureHandler((request, response, exception) -> {
-                            log.error("OAuth2 Login Failed", exception);
-                            response.sendRedirect("/auth/login?error");
-                        })
-                )
-                .logout(logout -> logout
-                        .logoutUrl("/logout")
-                        .logoutSuccessHandler(oAuth2LogoutSuccessHandler)
-                        .clearAuthentication(true)
-                        .invalidateHttpSession(true))
                 .addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
