--- src/main/java/com/devita/common/oauth/OAuth2LogoutSuccessHandler.java
@@ -1,59 +1,0 @@
-package com.devita.common.oauth;
-
-import com.devita.common.jwt.JwtTokenProvider;
-import com.devita.common.jwt.RefreshTokenService;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class OAuth2LogoutSuccessHandler implements LogoutSuccessHandler {
-
-    private final JwtTokenProvider jwtTokenProvider;
-    private final RefreshTokenService refreshTokenService;
-
-    @Override
-    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        log.info("Logout successful");
-        String refreshToken = null;
-        Cookie[] cookies = request.getCookies();
-        if (cookies != null) {
-            for (Cookie cookie : cookies) {
-                if ("refreshToken".equals(cookie.getName())) {
-                    refreshToken = cookie.getValue();
-                    break;
-                }
-            }
-        }
-
-        if (refreshToken != null && !refreshToken.isEmpty()) {
-            try {
-                Long userId = jwtTokenProvider.getUserIdFromRefreshToken(refreshToken);
-                refreshTokenService.deleteRefreshToken(userId);
-                log.info("Refresh token redis deleted");
-            } catch (Exception e) {
-                log.error("Error during logout process: {}", e.getMessage(), e);
-            }
-        }
-
-        Cookie refreshTokenCookie = new Cookie("refreshToken", null);
-        refreshTokenCookie.setHttpOnly(true);
-        refreshTokenCookie.setSecure(true);
-        refreshTokenCookie.setPath("/");
-        refreshTokenCookie.setMaxAge(0);
-        response.addCookie(refreshTokenCookie);
-        log.info("Refresh token cookie deleted");
-
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.getWriter().flush();
-    }
-}
