+++ src/main/java/com/devita/domain/user/domain/AuthDTO.java
@@ -1,0 +1,8 @@
+package com.devita.domain.user.domain;
+
+import lombok.Getter;
+
+@Getter
+public class AuthDTO {
+    String kakaoAccessToken;
+}
