+++ src/main/java/com/devita/domain/user/service/KakaoUserInfoService.java
@@ -1,0 +1,39 @@
+package com.devita.domain.user.service;
+
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.Map;
+
+@Service
+public class KakaoUserInfoService {
+    private static final String KAKAO_USER_INFO_URL = "https://kapi.kakao.com/v2/user/me";
+
+    public Map<String, Object> getUserInfo(String accessToken) {
+        RestTemplate restTemplate = new RestTemplate();
+
+        // Authorization 헤더 설정: "Bearer {accessToken}"
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "Bearer " + accessToken);
+
+        HttpEntity<?> entity = new HttpEntity<>(headers);
+
+
+        // Kakao API 호출
+        ResponseEntity<Map<String, Object>> response = restTemplate.exchange(
+                KAKAO_USER_INFO_URL,
+                HttpMethod.GET,
+                entity,
+                new ParameterizedTypeReference<>() {}
+        );
+
+        // 응답 바디에 사용자 정보가 담겨있음
+        return response.getBody();
+    }
+
+}
